Feat: [BADA-109] 메서드 자료형 수정에 따른 자료형 수정 #74

+++ src/main/java/com/TwoSeaU/BaData/domain/user/service/UserService.java
@@ -82,7 +82,7 @@ public class UserService {
 			.map(payment -> {
 				Post post = postRepository.findById(payment.getPost().getId())
 					.orElseThrow(() -> new GeneralException(TradeException.POST_NOT_FOUND));
-				Long postLikes = postLikesRepository.countByPostId(post.getId());
+				int postLikes = postLikesRepository.countByPostId(post.getId());
 
 				return GetPurchaseResponse.from(post, payment, postLikes);
 			})
